HBX-1692: Move class 'org.hibernate.tool.stat.BeanTableModel' to 'org.hibernate.tool.internal.util.BeanTableModel'
